fix: Move float parsing to character
- integers are parsed in `character` with their rust type names
- floats are parsed in `number` with their c typ names

This discrepancy was pointed out in Geal/nom#1506.

This moves floats to be in `character` and to use their rust type names
