+++ src/number/mod.rs
@@ -8,13 +8,10 @@ pub mod streaming;
 mod tests;
 
 use crate::error::ParseError;
-use crate::input::Compare;
-use crate::input::ParseTo;
 use crate::input::{
-  AsBytes, AsChar, InputIsStreaming, InputIter, InputLength, InputTake, InputTakeAtPosition,
-  IntoOutput, Offset, Slice,
+  AsBytes, AsChar, InputIsStreaming, InputIter, InputLength, InputTakeAtPosition, Slice,
 };
-use crate::lib::std::ops::{Range, RangeFrom, RangeTo};
+use crate::lib::std::ops::{RangeFrom, RangeTo};
 use crate::IResult;
 
 /// Configurable endianness
@@ -2196,224 +2193,5 @@ where
     streaming::hex_u32(input)
   } else {
     complete::hex_u32(input)
-  }
-}
-
-/// Recognizes floating point number in a byte string and returns the corresponding slice.
-///
-/// *Complete version*: Can parse until the end of input.
-///
-/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
-///
-/// # Example
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::recognize_float;
-///
-/// let parser = |s| {
-///   recognize_float(s)
-/// };
-///
-/// assert_eq!(parser("11e-1"), Ok(("", "11e-1")));
-/// assert_eq!(parser("123E-02"), Ok(("", "123E-02")));
-/// assert_eq!(parser("123K-01"), Ok(("K-01", "123")));
-/// assert_eq!(parser("abc"), Err(Err::Error(("abc", ErrorKind::Char))));
-/// ```
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::input::Streaming;
-/// use nom::number::recognize_float;
-///
-/// let parser = |s| {
-///   recognize_float(s)
-/// };
-///
-/// assert_eq!(parser(Streaming("11e-1;")), Ok((Streaming(";"), "11e-1")));
-/// assert_eq!(parser(Streaming("123E-02;")), Ok((Streaming(";"), "123E-02")));
-/// assert_eq!(parser(Streaming("123K-01")), Ok((Streaming("K-01"), "123")));
-/// assert_eq!(parser(Streaming("abc")), Err(Err::Error((Streaming("abc"), ErrorKind::Char))));
-/// ```
-#[inline(always)]
-pub fn recognize_float<T, E: ParseError<T>, const STREAMING: bool>(
-  input: T,
-) -> IResult<T, <T as IntoOutput>::Output, E>
-where
-  T: Slice<RangeFrom<usize>> + Slice<RangeTo<usize>>,
-  T: Clone + Offset,
-  T: InputIter + InputLength + InputIsStreaming<STREAMING>,
-  T: IntoOutput,
-  <T as InputIter>::Item: AsChar,
-  T: InputTakeAtPosition,
-  <T as InputTakeAtPosition>::Item: AsChar,
-{
-  if STREAMING {
-    streaming::recognize_float(input)
-  } else {
-    complete::recognize_float(input)
-  }
-}
-
-/// Recognizes a floating point number in text format
-///
-/// It returns a tuple of (`sign`, `integer part`, `fraction part` and `exponent`) of the input
-/// data.
-///
-/// *Complete version*: Can parse until the end of input.
-///
-/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
-///
-#[inline(always)]
-pub fn recognize_float_parts<T, E: ParseError<T>, const STREAMING: bool>(
-  input: T,
-) -> IResult<
-  T,
-  (
-    bool,
-    <T as IntoOutput>::Output,
-    <T as IntoOutput>::Output,
-    i32,
-  ),
-  E,
->
-where
-  T: Slice<RangeFrom<usize>> + Slice<RangeTo<usize>> + Slice<Range<usize>>,
-  T: Clone + Offset,
-  T: InputIter + InputTake + InputIsStreaming<STREAMING>,
-  T: IntoOutput,
-  <T as InputIter>::Item: AsChar + Copy,
-  T: InputTakeAtPosition + InputLength,
-  <T as InputTakeAtPosition>::Item: AsChar,
-  T: for<'a> Compare<&'a [u8]>,
-  T: AsBytes,
-{
-  if STREAMING {
-    streaming::recognize_float_parts(input)
-  } else {
-    complete::recognize_float_parts(input)
-  }
-}
-
-/// Recognizes floating point number in text format and returns a f32.
-///
-/// *Complete version*: Can parse until the end of input.
-///
-/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
-///
-/// # Example
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::float;
-///
-/// let parser = |s| {
-///   float(s)
-/// };
-///
-/// assert_eq!(parser("11e-1"), Ok(("", 1.1)));
-/// assert_eq!(parser("123E-02"), Ok(("", 1.23)));
-/// assert_eq!(parser("123K-01"), Ok(("K-01", 123.0)));
-/// assert_eq!(parser("abc"), Err(Err::Error(("abc", ErrorKind::Float))));
-/// ```
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// # use nom::input::Streaming;
-/// use nom::number::float;
-///
-/// let parser = |s| {
-///   float(s)
-/// };
-///
-/// assert_eq!(parser(Streaming("11e-1 ")), Ok((Streaming(" "), 1.1)));
-/// assert_eq!(parser(Streaming("11e-1")), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(parser(Streaming("123E-02")), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(parser(Streaming("123K-01")), Ok((Streaming("K-01"), 123.0)));
-/// assert_eq!(parser(Streaming("abc")), Err(Err::Error((Streaming("abc"), ErrorKind::Float))));
-/// ```
-#[inline(always)]
-pub fn float<T, E: ParseError<T>, const STREAMING: bool>(input: T) -> IResult<T, f32, E>
-where
-  T: Slice<RangeFrom<usize>> + Slice<RangeTo<usize>> + Slice<Range<usize>>,
-  T: Clone + Offset + Compare<&'static str>,
-  T: InputIter + InputLength + InputTake + InputIsStreaming<STREAMING>,
-  T: IntoOutput,
-  <T as IntoOutput>::Output: ParseTo<f32>,
-  <T as InputIter>::Item: AsChar + Copy,
-  <T as InputIter>::IterElem: Clone,
-  T: InputTakeAtPosition,
-  <T as InputTakeAtPosition>::Item: AsChar,
-  T: AsBytes,
-  T: for<'a> Compare<&'a [u8]>,
-{
-  if STREAMING {
-    streaming::float(input)
-  } else {
-    complete::float(input)
-  }
-}
-
-/// Recognizes floating point number in text format and returns a f64.
-///
-/// *Complete version*: Can parse until the end of input.
-///
-/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
-///
-/// # Example
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::double;
-///
-/// let parser = |s| {
-///   double(s)
-/// };
-///
-/// assert_eq!(parser("11e-1"), Ok(("", 1.1)));
-/// assert_eq!(parser("123E-02"), Ok(("", 1.23)));
-/// assert_eq!(parser("123K-01"), Ok(("K-01", 123.0)));
-/// assert_eq!(parser("abc"), Err(Err::Error(("abc", ErrorKind::Float))));
-/// ```
-///
-/// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// # use nom::input::Streaming;
-/// use nom::number::double;
-///
-/// let parser = |s| {
-///   double(s)
-/// };
-///
-/// assert_eq!(parser(Streaming("11e-1 ")), Ok((Streaming(" "), 1.1)));
-/// assert_eq!(parser(Streaming("11e-1")), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(parser(Streaming("123E-02")), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(parser(Streaming("123K-01")), Ok((Streaming("K-01"), 123.0)));
-/// assert_eq!(parser(Streaming("abc")), Err(Err::Error((Streaming("abc"), ErrorKind::Float))));
-/// ```
-#[inline(always)]
-pub fn double<T, E: ParseError<T>, const STREAMING: bool>(input: T) -> IResult<T, f64, E>
-where
-  T: Slice<RangeFrom<usize>> + Slice<RangeTo<usize>> + Slice<Range<usize>>,
-  T: Clone + Offset + Compare<&'static str>,
-  T: InputIter + InputLength + InputTake + InputIsStreaming<STREAMING>,
-  T: IntoOutput,
-  <T as IntoOutput>::Output: ParseTo<f64>,
-  <T as InputIter>::Item: AsChar + Copy,
-  <T as InputIter>::IterElem: Clone,
-  T: InputTakeAtPosition,
-  <T as InputTakeAtPosition>::Item: AsChar,
-  T: AsBytes,
-  T: for<'a> Compare<&'a [u8]>,
-{
-  if STREAMING {
-    streaming::double(input)
-  } else {
-    complete::double(input)
   }
 }
